Refactor streaming functions into Request
* extracting body still not quite working correctly for streams
